src/main/webapp/app/entities/perfil-acesso/delete/perfil-acesso-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { PerfilAcessoService } from '../service/perfil-acesso.service';

import { PerfilAcessoDeleteDialogComponent } from './perfil-acesso-delete-dialog.component';

describe('PerfilAcesso Management Delete Component', () => {
  let comp: PerfilAcessoDeleteDialogComponent;
  let fixture: ComponentFixture<PerfilAcessoDeleteDialogComponent>;
  let service: PerfilAcessoService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [PerfilAcessoDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(PerfilAcessoDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(PerfilAcessoDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(PerfilAcessoService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
